Reduce database locking
Instead of doing the entire database update in one transaction
use the new batch function to batch discovery and brute force results
in groups of 50 or 2 seconds.  This way a transaction is never open for
very long, and the scan is checkpointed every 2 seconds in case it is
control-c'd

This looks like a big change but it's mostly indentation.